+++ src/pytanis/pretalx/__init__.py
@@ -1,5 +1,6 @@
 """Functionality around the Pretalx API"""
 
-from .api import PretalxAPI
+from .client import PretalxClient
+from .utils import reviews_as_df, speakers_as_df, subs_as_df
 
-__all__ = ["PretalxAPI"]
+__all__ = ["PretalxClient", "subs_as_df", "speakers_as_df", "reviews_as_df"]