+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - use [gspread](https://docs.gspread.org/) to handle the low-level GoogleAPI
 - use [gspread-dataframe](https://gspread-dataframe.readthedocs.io/) to handle converting a worksheet into a dataframe
 - timeout of 60s for PretalxAPI as it is really slow which caused a lot of timeout errors
+- rename `*API` to `*Client` as it's rather a client for an API
+- moved some functionality from `review` to `pretalx.utils`
 
 ## Version 0.1.1 (2023-01-16)